Remove split package situation
