+++ project-api/ballerina-projects/src/main/java/io/ballerina/projects/directory/ProjectLoader.java
@@ -15,14 +15,12 @@
  *  specific language governing permissions and limitations
  *  under the License.
  */
-package io.ballerina.projects;
+package io.ballerina.projects.directory;
 
 import io.ballerina.build.DocumentId;
 import io.ballerina.build.Module;
 import io.ballerina.build.ModuleId;
 import io.ballerina.build.Project;
-import io.ballerina.projects.directory.BuildProject;
-import io.ballerina.projects.directory.SingleFileProject;
 import io.ballerina.projects.utils.ProjectConstants;
 
 import java.nio.file.Path;
