+++ project-api/ballerina-projects/src/test/java/io/ballerina/projects/test/TestBuildProject.java
@@ -29,8 +29,8 @@ import io.ballerina.build.ModuleName;
 import io.ballerina.build.Package;
 import io.ballerina.build.PackageCompilation;
 import io.ballerina.build.PackageId;
-import io.ballerina.projects.ProjectLoader;
 import io.ballerina.projects.directory.BuildProject;
+import io.ballerina.projects.directory.ProjectLoader;
 import io.ballerina.projects.utils.ProjectConstants;
 import io.ballerina.projects.utils.ProjectUtils;
 import org.testng.Assert;
