+++ project-api/ballerina-projects/src/test/java/io/ballerina/projects/test/TestUnknownTypeProject.java
@@ -22,8 +22,8 @@ import io.ballerina.build.Module;
 import io.ballerina.build.ModuleId;
 import io.ballerina.build.Package;
 import io.ballerina.build.Project;
-import io.ballerina.projects.ProjectLoader;
 import io.ballerina.projects.directory.BuildProject;
+import io.ballerina.projects.directory.ProjectLoader;
 import io.ballerina.projects.directory.SingleFileProject;
 import org.testng.Assert;
 import org.testng.annotations.Test;
